// src/utils/types/types.ts
export type TCryptoValue = {
  0: {
    name: string;
    rateTo: {
      'USDT'?: number;
      'USD'?: number;
      'BTC'?: number;
      'ETH'?: number;
      'EUR'?: number;
      'RUB'?: number;
    };
  };
  1: {
    name: string;
    rateTo: {
      'USDT'?: number;
      'USD'?: number;
      'BTC'?: number;
      'ETH'?: number;
      'EUR'?: number;
      'RUB'?: number;
    };
  };
  2: {
    name: string;
    rateTo: {
      'USDT'?: number;
      'USD'?: number;
      'BTC'?: number;
      'ETH'?: number;
      'EUR'?: number;
      'RUB'?: number;
    };
  };
};

export type TDefaultValue = {
  0: {
    name: string;
    rateTo: {
      'USDT': number;
      'BTC': number;
      'ETH': number;
    };
  };
  1: {
    name: string;
    rateTo: {
        'USDT': number;
        'BTC': number;
        'ETH': number;
      };
  };
  2: {
    name: string;
    rateTo: {
        'USDT': number;
        'BTC': number;
        'ETH': number;
      };
  };
};